+++ app/components/Analyst/RFI/RFI.tsx
@@ -107,7 +107,7 @@ const RFI: React.FC<Props> = ({ rfiDataByRfiDataId, id }) => {
         formData={jsonData}
         noValidate
         tagName="div"
-        formContext={{ applicationId, rfiId: rowId }}
+        formContext={{ applicationId, rfiId: rowId, skipUnsavedWarning: true }}
         // Pass children to hide submit button
         // eslint-disable-next-line react/no-children-prop
         children