+++ app/tests/pages/analyst/dashboard.test.tsx
@@ -34,6 +34,7 @@ const mockQueryPayload = {
               zones: [1, 2],
               program: 'CCBC',
               status: 'received',
+              package: '1',
               applicationFormTemplate9DataByApplicationId: {
                 nodes: [
                   {
@@ -64,6 +65,7 @@ const mockQueryPayload = {
               zones: [],
               program: 'CCBC',
               status: 'approved',
+              package: null,
               applicationFormTemplate9DataByApplicationId: {
                 nodes: [
                   {
@@ -853,6 +855,62 @@ describe('The index page', () => {
     });
   });
 
+  it('global filter correctly filters communities and expand the row with match highlighting', async () => {
+    jest
+      .spyOn(moduleApi, 'useFeature')
+      .mockReturnValue(mockShowCbcProjects(true));
+
+    pageTestingHelper.loadQuery();
+    pageTestingHelper.renderPage();
+
+    expect(screen.getByText('CCBC-010001')).toBeInTheDocument();
+    expect(screen.getByText('CCBC-010002')).toBeInTheDocument();
+
+    const globalSearch = screen.getByPlaceholderText('Search');
+    expect(globalSearch).toBeInTheDocument();
+
+    fireEvent.change(globalSearch, {
+      target: { value: 'Bear Lake' },
+    });
+
+    await waitFor(() => {
+      expect(screen.queryByText('CCBC-010001')).not.toBeInTheDocument();
+      expect(screen.getByText('CCBC-010002')).toBeInTheDocument();
+
+      expect(screen.getByText('Bear Lake')).toHaveStyle(
+        'background-color: #FCBA19'
+      );
+    });
+  });
+
+  it('clear filters correctly clears global filter and restore data', async () => {
+    jest
+      .spyOn(moduleApi, 'useFeature')
+      .mockReturnValue(mockShowCbcProjects(true));
+
+    pageTestingHelper.loadQuery();
+    pageTestingHelper.renderPage();
+
+    expect(screen.getByText('5555')).toBeInTheDocument();
+    const globalSearch = screen.getByPlaceholderText('Search');
+    expect(globalSearch).toBeInTheDocument();
+
+    fireEvent.change(globalSearch, {
+      target: { value: 'Bear Lake' },
+    });
+
+    await waitFor(() => {
+      expect(screen.queryByText('CCBC-010001')).not.toBeInTheDocument();
+    });
+
+    const clearFiltersBtn = screen.getByText('Clear Filtering');
+    await act(async () => {
+      fireEvent.click(clearFiltersBtn);
+    });
+
+    expect(screen.queryByText('CCBC-010001')).toBeInTheDocument();
+  });
+
   it('cbc statuses are duplicated for both analyst and external statuses', async () => {
     jest
       .spyOn(moduleApi, 'useFeature')
@@ -873,6 +931,44 @@ describe('The index page', () => {
     pageTestingHelper.renderPage();
 
     expect(screen.getAllByText('Reporting complete')).toHaveLength(6);
+  });
+
+  it('should correctly filter by package filter', async () => {
+    jest
+      .spyOn(moduleApi, 'useFeature')
+      .mockReturnValue(mockShowCbcProjects(true));
+
+    pageTestingHelper.loadQuery();
+    pageTestingHelper.renderPage();
+
+    expect(screen.getByText('CCBC-010001')).toBeVisible();
+    expect(screen.getByText('CCBC-010002')).toBeVisible();
+
+    const columnActions = document.querySelectorAll(
+      '[aria-label="Show/Hide filters"]'
+    )[0];
+
+    await act(async () => {
+      fireEvent.click(columnActions);
+    });
+
+    const packageFilter = screen.getAllByText('Filter by Package')[0];
+
+    expect(packageFilter).toBeInTheDocument();
+
+    await act(async () => {
+      fireEvent.keyDown(packageFilter, { key: 'Enter', code: 'Enter' });
+    });
+
+    const option = screen.getByRole('option', { name: 'Unassigned' });
+    await act(async () => {
+      fireEvent.click(option);
+    });
+
+    await waitFor(() => {
+      expect(screen.getByText('CCBC-010002')).toBeInTheDocument();
+      expect(screen.queryByText('CCBC-010001')).not.toBeInTheDocument();
+    });
   });
 
   it('should correctly filter the cbc projects by analyst status filter', async () => {